# backend-arqtool-main/ViewModels/EditUsuarioViewModel.cs
namespace caiobadev_api_arqtool.ViewModels {
    public class EditUsuarioViewModel {
        public string Id { get; set; }
        public string Nome { get; set; }
        public string Sobrenome { get; set; }
        public DateTime DataNascimento { get; set; }
        public string Telefone { get; set; }
    }
}