# backend-arqtool-main/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ApiArqTool",
    "version": "v1"
  },
  "paths": {
...
    "/api/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuarioInfo"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EditUsuarioViewModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EditUsuarioViewModel": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "sobrenome": {
          "type": "string"
        },
        "dataNascimento": {
          "format": "date-time",
          "type": "string"
        },
        "telefone": {
          "type": "string"
        }
      }
    },
...
    "UsuarioInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "sobrenome": {
          "type": "string"
        },
        "dataNascimento": {
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "perfil": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Header de autorização JWT usando o esquema Bearer.\r\n\r\nInforme 'Bearer'[espaço] e o seu token.\r\n\r\nExemplo: 'Bearer 12345abcdef' "
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}